+++ expenses/apps.py
@@ -3,3 +3,6 @@ from django.apps import AppConfig
 
 class ExpensesConfig(AppConfig):
     name = 'expenses'
+
+    def ready(self):
+        from . import signals  # noqa: F401